refactor(presets): Rename to *.preset.ts and enforce sort-keys via oxlint
