+++ lib/config/presets/internal/default.preset.ts
@@ -1,6 +1,5 @@
 import type { Preset } from '../types.ts';
 
-/* eslint sort-keys: ["error", "asc", {caseSensitive: false, natural: true}] */
 export const presets: Record<string, Preset> = {
   approveMajorUpdates: {
     description: 'Require Dependency Dashboard approval for `major` updates.',
