+++ lib/config/presets/internal/global.preset.ts
@@ -1,7 +1,5 @@
 import type { GlobalPreset } from '../types.ts';
 
-/* eslint sort-keys: ["error", "asc", {caseSensitive: false, natural: true}] */
-
 export const presets: Record<string, GlobalPreset> = {
   safeEnv: {
     allowedEnv: ['GO*', 'GRADLE_OPTS', 'RUSTC_BOOTSTRAP'],
