+++ lib/workers/repository/process/lookup/index.spec.ts
@@ -1,7 +1,7 @@
 import { codeBlock } from 'common-tags';
 import * as hostRules from '../../../../../lib/util/host-rules.ts';
 import { getConfig } from '../../../../config/defaults.ts';
-import { supportedDatasources as presetSupportedDatasources } from '../../../../config/presets/internal/merge-confidence.ts';
+import { supportedDatasources as presetSupportedDatasources } from '../../../../config/presets/internal/merge-confidence.preset.ts';
 import type { AllConfig } from '../../../../config/types.ts';
 import { CONFIG_VALIDATION } from '../../../../constants/error-messages.ts';
 import { CustomDatasource } from '../../../../modules/datasource/custom/index.ts';
